+++ AutoBlogger/.gitignore
@@ -3,3 +3,4 @@
 __pycache__
 .chroma
 logs
+markdowns/
