+++ AutoBlogger/README.md
@@ -1,2 +1,11 @@
+only content creation
+python .\server_src\content_creation\test.py
+
+mcp 
 run each server individually in terminal
+python .\mcp_files\servers\content_creation_server.py
+python .\mcp_files\servers\hashnode_server.py
+python .\mcp_files\servers\markdown_file_server.py
+
 run client
+python .\mcp_files\client\client.py
