+++ AutoBlogger/mcp_files/servers/markdown_file_server.py
@@ -1,3 +1,6 @@
+import sys
+import os
+sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), "../..")))
 from logging_config import get_logger
 
 logger = get_logger("MarkdownFileServer")
