+++ AutoBlogger/server_src/content_creation/graph/nodes/extract_links_content.py
@@ -1,7 +1,7 @@
 from logging_config import get_logger
 
 logger = get_logger("ContentCreationServer")
-from content_creation.graph.state import ContentState
+from server_src.content_creation.graph.state import ContentState
 from typing import Dict
 import trafilatura
