+++ AutoBlogger/server_src/content_creation/test.py
@@ -1,8 +1,10 @@
 from dotenv import load_dotenv
 load_dotenv()
-
-from content_creation.graph.graph import content_graph
-from content_creation.utils import draw_stylish_graph
+import os
+import sys
+sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), "../..")))
+from server_src.content_creation.graph.graph import content_graph
+from server_src.content_creation.utils import draw_stylish_graph
 
 if __name__=='__main__':
     draw_stylish_graph(content_graph, output_name="content_graph")
